Added a readme with some style guidelines.
Implemented savehist and modified the return values to be more usable.

I've gone through cobyla.py and it seems pretty close to the original.

Remaining work is to implement retmsg and to add in all pre/post conditions
and retrofit existing pre/post conditions to match the style guide.

Then we're ready for a deeper conversation on whether any of this is useful XD